refactor: move Coordinates value object out of offers.services to break import cycle
