+++ shargain/offers/geo.py
@@ -1,0 +1,6 @@
+from typing import NamedTuple
+
+
+class Coordinates(NamedTuple):
+    lat: float
+    lon: float
